配合auth的認證安裝Sanctum與相關設定；修復新建文章的預覽－>重寫BUG，增加確認儲存圖片是否已存在；調整文章編輯頁面的排版；在文章一覽頁面強調未完成文章

+++ app/Http/Controllers/api/PostController.php
@@ -274,13 +274,17 @@ class PostController extends Controller
                     if (isset($pics) && gettype($pics) == 'array') { //確認前台有圖片時才執行
                         foreach ($pics as $pic) { //將文章內有的圖檔路徑一一取出
                             if ($tempName == $pic) { //確定該暫存圖有放在文章內時
-                                $result = Storage::move($value, str_replace($tempPath, $path, $value)); //移動檔案到正式儲存位置
+                                if (!Storage::exists($value)) {
+                                    $result = Storage::move($value, str_replace($tempPath, $path, $value)); //移動檔案到正式儲存位置
+                                }
                                 break; //結束，後面不用跑了
                             }
                         }
                     }
                     if (!$result) { //若圖片移動失敗時報錯
-                        $newPost->delete();
+                        if ($action != 'rewrite') {
+                            $newPost->delete();
+                        }
                         return $this->makeJson(0, $result, 'MOVE_FILE_ERROR');
                     }
                 }
@@ -291,7 +295,9 @@ class PostController extends Controller
                     $old = Post::Where('id', $request->postId)->get()->first(); //從資料庫中找出修改前的舊文章
                     $result = $old->update(['state' => 0]); //修改文章狀態為凍結，不刪除資料
                     if (!$result) { //修改狀態失敗時報錯
-                        $newPost->delete();
+                        if ($action != 'rewrite') {
+                            $newPost->delete();
+                        }
                         return $this->makeJson(0, $result, 'FREEZE_OLDER_ERROR');
                     }
                 }
@@ -305,10 +311,15 @@ class PostController extends Controller
                 Cache::put(Auth::id(), $tags);
                 if ($request->has('image') && !is_null($request->image) && $request->image != '/storage/post/default/default.jpg') {
                     $saveImage = str_replace('temp/' . Auth::id(), 'post/' . $id, $request->image);
-                    $image = str_replace('storage', 'public', $request->image); //取出文章代表圖像
-                    $result = Storage::move($image, str_replace($tempPath, $path, $image)); //移動檔案到正式儲存位置
-                    if (!$result) {
-                        return $this->makeJson(0, $result, 'IMAGE_SAVE_ERROR');
+                    // $check = Storage::exists(str_replace('storage', 'public', $request->image));
+                    $check = str_replace('storage', 'public', $request->image);
+                    // return $this->makeJson(0, $check, 'HE');
+                    if (!Storage::exists($check)) {
+                        $image = str_replace('storage', 'public', $request->image); //取出文章代表圖像
+                        $result = Storage::move($image, str_replace($tempPath, $path, $image)); //移動檔案到正式儲存位置
+                        if (!$result) {
+                            return $this->makeJson(0, $result, 'IMAGE_SAVE_ERROR');
+                        }
                     }
                     $result = Storage::delete(Storage::allFiles($tempPath)); //文章內所有的圖片儲存完畢後，資料夾內剩餘的檔案
                     if (!$result) { //刪除剩餘圖片失敗時報錯
@@ -326,7 +337,9 @@ class PostController extends Controller
             }
         } catch (\Throwable $ex) {
             // return $this->makeJson(0, ['Message' => $ex->getMessage(), 'Code' => $ex->getCode(), 'Line' => $ex->getLine()], $ex->getMessage()); //程式碼錯誤
-            Post::where('id', $id)->get()->first()->delete();
+            if ($action != 'rewrite') {
+                Post::where('id', $id)->get()->first()->delete();
+            }
             return $this->makeJson(0, ['Message' => $ex], 'ERROR'); //程式碼錯誤
         }
